+++ src/main/java/com/example/demo/repository/CompliantRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CompliantRepository extends JpaRepository<Compliant, Integer> {
+public interface CompliantRepository extends JpaRepository<Compliant, Long> {
     Optional<Compliant> findByReferenceNumber(String referenceNumber);
 }
