+++ src/main/java/com/example/demo/repository/FileRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Files;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<Files,Integer> {
+public interface FileRepository extends JpaRepository<Files,Long> {
 }
