+++ src/main/java/com/example/demo/repository/OrganizationalUnitRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.OrganizationalUnit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizationalUnitRepository extends JpaRepository<OrganizationalUnit,Integer> {
+public interface OrganizationalUnitRepository extends JpaRepository<OrganizationalUnit,Long> {
 }
